[MOV] test_mail: move access tests in their own file
Purpose is to separate tests about content from tests about access. It
eases bugfixing and improving features while keeping other tests stable
through versions.

Task-4320561
