+++ addons/mail/tests/test_res_users.py
@@ -3,7 +3,9 @@
 
 from psycopg2 import IntegrityError
 
+from odoo.addons.base.tests.common import HttpCaseWithUserDemo
 from odoo.addons.mail.tests.common import MailCommon, mail_new_test_user
+from odoo.tests import tagged
 from odoo.tools import mute_logger
 
 
@@ -20,3 +22,19 @@ class TestUser(MailCommon):
                 notification_type='inbox',
                 groups='base.group_portal',
             )
+
+
+@tagged('-at_install', 'post_install')
+class TestUserTours(HttpCaseWithUserDemo):
+
+    def test_user_modify_own_profile(self):
+        """" A user should be able to modify their own profile.
+        Even if that user does not have access rights to write on the res.users model. """
+        if 'hr.employee' in self.env and not self.user_demo.employee_id:
+            self.env['hr.employee'].create({
+                'name': 'Marc Demo',
+                'user_id': self.user_demo.id,
+            })
+        self.user_demo.tz = "Europe/Brussels"
+        self.start_tour("/web", "mail/static/tests/tours/user_modify_own_profile_tour.js", login="demo")
+        self.assertEqual(self.user_demo.email, "updatedemail@example.com")
