+++ addons/test_mail/tests/__init__.py
@@ -7,6 +7,7 @@ from . import test_mail_composer
 from . import test_mail_composer_mixin
 from . import test_mail_followers
 from . import test_mail_message
+from . import test_mail_message_security
 from . import test_mail_mail
 from . import test_mail_gateway
 from . import test_mail_multicompany
